Uses PRIx64 and PRIu64 for portable lu and ll print format

+++ picoquictest/intformattest.c
@@ -253,8 +253,8 @@ int varint_test()
                 } else if (length == 0) {
                     continue;
                 } else if (n64 != test->decoded) {
-                    fprintf(stderr, "Varint: unexpected value %llu [expected %llu]", 
-                        (unsigned long long)n64, (unsigned long long)test->decoded);
+                    fprintf(stderr, "Varint: unexpected value %" PRIu64 " [expected %" PRIu64 "]",
+                        n64, test->decoded);
                     test_ret = -1;
                 } else if (test->is_canonical != 0) {
                     uint8_t encoding[8];
